Extract custom assertions in their own file
